% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{Takes a hex-colour as argument and is able to add an alpha value (opacity) to it}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{Any color that can be transformed to RGB using \code{col2rgb}.}

\item{alpha}{The alpha value we wish to assign to the color.}
}
\description{
Takes a hex-colour as argument and is able to add an alpha value (opacity) to it
}
