% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOptim.R
\name{BayesOptim}
\alias{BayesOptim}
\title{Minimizing objective functions with Gaussian processes}
\usage{
BayesOptim(
  FUN,
  parameters,
  kappa = 3,
  n = 3^length(parameters),
  maxiter = 100,
  convergence = 5
)
}
\arguments{
\item{FUN}{The objective function to minimise}

\item{parameters}{A list of parameters}

\item{kappa}{Multiplier parameter to calculate the acquisition function. Higher values of \code{kappa} will trigger more exploration on the cost of exploitation.}

\item{n}{Number of randomly chosen points to evaluate in the beginning.}

\item{maxiter}{Maximum number of points to evaluate.}

\item{convergence}{Stopping condition. If the same point is chosen as the minimum for \code{convergence} times, the algorithm stops.}
}
\description{
Function to fit a Gaussian process regression model to minimize an objective function.
Bayesian optimization techniques are most useful if the evaluation of the obejctive functions are costly.
}
\examples{
fn <- function(x) (x[1]-5)^2 + (x[2]-5)^2 + sin(12*x[1]) + sin(9*x[1])  - cos(17*x[2]) + sin(13*x[2]*x[1])*cos(x[2]) + cos(11*x[2]) + sin(23*x[2])*cos(21*x[1])
parameters <- list(seq(3,7,length=50), seq(0,10,length=50))
results <- BayesOptim(FUN = fn, parameters = parameters)

}
