% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkmode.R
\name{darkmode}
\alias{darkmode}
\title{Set graphical parameters to match RStudio dark mode}
\usage{
darkmode()
}
\description{
This function sets the graphical parameters so they match the RStudio dark mode.
}
