% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_colors}
\alias{get_colors}
\title{Get custom sets of colors}
\usage{
get_colors(type = "darkmode")
}
\description{
This function return a list with colors. For now, this is only the list darkmode, which contains the colors used in the RStudio darkmode design.
}
