% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalizing a vector}
\usage{
normalize(x, na.rm = TRUE)
}
\description{
This function normalizes an input vector \code{x}, i.e. it re-scales the vector between 0 and 1.
}
