% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_vs_obs.R
\name{pred_vs_obs}
\alias{pred_vs_obs}
\title{Plotting predicted vs. tested values}
\usage{
pred_vs_obs(
  predicted,
  observed,
  main = "",
  fit.col = "cadetblue",
  plot = FALSE
)
}
\description{
Plotting predicted vs. tested values
}
