% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Compute the standard error}
\usage{
se(x, ...)
}
\arguments{
\item{x}{A numeric vector of which we wish to compute the standard error.}

\item{...}{Additional arguments such as for example \code{na.rm}.}
}
\description{
Compute the standard error
}
