% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Smooth hyperspectral data}
\usage{
smoothing(x, lambda_in, lambda_out = lambda_in, span = 0.1, prefix = "rho_")
}
\arguments{
\item{x}{A data frame or matrix whose rows contain observations.}

\item{lambda_in}{A vector of wavelengths (one wavelength for every column in x).}

\item{lambda_out}{A vector of desired wavelengths for the interpolation.}

\item{span}{A control parameter for the smoothness of the interpolation.}
}
\description{
Smooth hyperspectral data
}
