% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{Softmax}
\usage{
softmax(x, ...)
}
\arguments{
\item{x}{A numeric vector of which we wish to compute the softmax.}

\item{...}{Additional arguments such as for example \code{na.rm}.}
}
\description{
Compute the smooth approximation to the maximum function (called "softmax" a.k.a. normalized exponential function).
}
