% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VI.R
\name{VI}
\alias{VI}
\title{Compute vegetation indices}
\usage{
VI(x, lambda)
}
\arguments{
\item{x}{A matrix or data frame with hyperspectral observations.}

\item{lambda}{A vector of wavelengths}
}
\description{
Compute vegetation indices
}
